requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
email_validator==2.2.0
Faker==37.1.0
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
